# backend/internal/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Basis Social API",
    "version": "1.0.0-alpha",
    "description": "Stable REST and world-server integration contract for BasisVR. ActivityPub endpoints are intentionally outside this client API contract."
  },
  "servers": [{"url": "/"}],
...
  "paths": {
...
    "/api/v1/events": {
      "get": {"tags":["Events"],"operationId":"listEvents","parameters":[{"$ref":"#/components/parameters/Limit"},{"$ref":"#/components/parameters/Cursor"}],"responses":{"200":{"description":"Cursor page of public events","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EventPage"}}}},"default":{"$ref":"#/components/responses/Error"}}},
      "post": {"tags":["Events"],"operationId":"createEvent","security":[{"bearerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EventWrite"}}}},"responses":{"201":{"description":"Created event","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Event"}}}},"default":{"$ref":"#/components/responses/Error"}}}
    },
    "/api/v1/events/{eventRef}": {"get": {"tags":["Events"],"operationId":"getEvent","parameters":[{"$ref":"#/components/parameters/EventRef"}],"responses":{"200":{"description":"Event; GET resolves eventRef as slug.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Event"}}}},"default":{"$ref":"#/components/responses/Error"}}},"patch": {"tags":["Events"],"operationId":"updateEvent","security":[{"bearerAuth":[]}],"parameters":[{"$ref":"#/components/parameters/EventRef"}],"responses":{"200":{"description":"Updated event; PATCH resolves eventRef as UUID."},"default":{"$ref":"#/components/responses/Error"}}},"delete":{"tags":["Events"],"operationId":"deleteEvent","security":[{"bearerAuth":[]}],"parameters":[{"$ref":"#/components/parameters/EventRef"}],"responses":{"204":{"description":"Deleted; DELETE resolves eventRef as UUID."},"default":{"$ref":"#/components/responses/Error"}}}},
    "/api/v1/events/{id}/rsvp": {"post": {"tags":["Events"],"operationId":"rsvpEvent","security":[{"bearerAuth":[]}],"parameters":[{"$ref":"#/components/parameters/ID"}],"responses":{"200":{"description":"RSVP updated"},"default":{"$ref":"#/components/responses/Error"}}}},
    "/api/v1/events/{id}/announce": {"post": {"tags":["Events"],"operationId":"announceEvent","security":[{"bearerAuth":[]}],"parameters":[{"$ref":"#/components/parameters/ID"}],"responses":{"202":{"description":"Announcement accepted"},"default":{"$ref":"#/components/responses/Error"}}}},
...
    "/api/v1/groups/{id}/events": {"get":{"tags":["Groups"],"operationId":"listGroupEvents","parameters":[{"$ref":"#/components/parameters/ID"}],"responses":{"200":{"description":"Linked events"},"default":{"$ref":"#/components/responses/Error"}}},"post":{"tags":["Groups"],"operationId":"addGroupEvent","security":[{"bearerAuth":[]}],"parameters":[{"$ref":"#/components/parameters/ID"}],"responses":{"204":{"description":"Event linked"},"default":{"$ref":"#/components/responses/Error"}}}},
    "/api/v1/groups/{id}/events/{eventId}": {"delete":{"tags":["Groups"],"operationId":"removeGroupEvent","security":[{"bearerAuth":[]}],"parameters":[{"$ref":"#/components/parameters/ID"},{"name":"eventId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"204":{"description":"Event unlinked"},"default":{"$ref":"#/components/responses/Error"}}}},
...
    "/api/v1/realtime/events": {"get":{"tags":["Realtime"],"operationId":"streamRealtimeEvents","security":[{"bearerAuth":[]}],"parameters":[{"name":"Last-Event-ID","in":"header","description":"Redis Stream cursor from the last processed event. Omit for a live-only connection.","schema":{"type":"string","pattern":"^[0-9]+-[0-9]+$"}}],"responses":{"200":{"description":"Server-Sent Events stream with optional durable replay","content":{"text/event-stream":{"schema":{"type":"string"}}}},"default":{"$ref":"#/components/responses/Error"}}}},
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {"type":"http","scheme":"bearer","bearerFormat":"JWT"},
      "serviceToken": {"type":"apiKey","in":"header","name":"X-Basis-Service-Token"}
    },
    "parameters": {
      "ID": {"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},
...
      "EventRef": {"name":"eventRef","in":"path","required":true,"schema":{"type":"string"}},
      "Limit": {"name":"limit","in":"query","schema":{"type":"integer","minimum":1,"maximum":100,"default":24}},
      "Cursor": {"name":"cursor","in":"query","description":"Opaque cursor returned by the previous page.","schema":{"type":"string"}}
    },
    "responses": {
      "Error": {"description":"Error response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}}
    },
...
    "schemas": {
      "Error": {"type":"object","required":["error"],"properties":{"error":{"type":"object","required":["code","message"],"properties":{"code":{"type":"string"},"message":{"type":"string"}}}}},
...
      "Owner": {"type":"object","required":["actorId","acct","displayName"],"properties":{"actorId":{"type":"string","format":"uuid"},"acct":{"type":"string"},"displayName":{"type":"string"}}},
      "Pagination": {"type":"object","required":["nextCursor","limit"],"properties":{"nextCursor":{"type":["string","null"]},"limit":{"type":"integer","minimum":1,"maximum":100}}},
...
      "Event": {"type":"object","required":["id","slug","name","owner","world","startTime","endTime","visibility"],"properties":{"id":{"type":"string","format":"uuid"},"slug":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"},"owner":{"$ref":"#/components/schemas/Owner"},"world":{"type":"object"},"startTime":{"type":"string","format":"date-time"},"endTime":{"type":"string","format":"date-time"},"launchUrl":{"type":"string"},"visibility":{"type":"string"},"metadata":{"type":"object"}}},
      "EventWrite": {"type":"object","required":["worldId","name","startTime","endTime"],"properties":{"worldId":{"type":"string","format":"uuid"},"slug":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"},"startTime":{"type":"string","format":"date-time"},"endTime":{"type":"string","format":"date-time"},"visibility":{"type":"string"},"launchUrl":{"type":"string"},"metadata":{"type":"object"}}},
      "EventPage": {"type":"object","required":["data","pagination"],"properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/Event"}},"pagination":{"$ref":"#/components/schemas/Pagination"}}},
...
    }
  }
}